Mac: events works when CC_DIRECTOR_MAC_USE_DISPLAY_LINK_THREAD is enabled
Events work correctly when CC_DIRECTOR_MAC_USE_DISPLAY_LINK_THREAD is
enabled.
They are dispatched by the display link thread.


Former-commit-id: 15ed2bc79b74bb13dcd8022d8336a54ede94140c

+++ cocos2d/CCConfiguration.h
@@ -57,7 +57,7 @@ enum {
 	BOOL			supportsNPOT_;
 	BOOL			supportsBGRA8888_;
 	BOOL			supportsDiscardFramebuffer_;
-	NSUInteger		OSVersion_;
+	unsigned int	OSVersion_;
 	GLint			maxSamplesAllowed_;
 }
 
@@ -97,7 +97,7 @@ enum {
  
  @since v0.99.5
  */
-@property (nonatomic, readonly) NSUInteger OSVersion;
+@property (nonatomic, readonly) unsigned int OSVersion;
 
 /** returns a shared instance of the CCConfiguration */
 +(CCConfiguration *) sharedConfiguration;
